Add files via upload
There are two types of port numbers - Local and global

Local Port Number:
When a socket binds to a port to establish a network connection, it is assigned a local port number.
This local port number helps identify the socket on the local machine.
For a client, the local port number is assigned by the operating system when it initiates a connection to a server. For a server, the local port number is typically predetermined and specified in the server's code.
The local port number is used along with the IP address of the local machine to uniquely identify the communication endpoint for the socket.
It's important to note that the local port number is relevant only within the context of the local machine and is not visible to remote machines on the network.

Port (in Networking):
In networking, a port is a logical construct used to identify different types of communication endpoints on a networked device (usually identified by an IP address).
Ports are represented by numbers and are used to differentiate between different services or applications running on the same device.
Ports are categorized into three ranges: well-known ports, registered ports, and dynamic or private ports, each serving different purposes.
Ports are used in conjunction with IP addresses to establish network connections between devices.
Unlike local port numbers, which are specific to a single machine, ports are used universally to route traffic between devices on a network.

Key Differences:

The local port number is specific to a particular socket on a local machine and helps identify the socket within the local operating system.
Ports, on the other hand, are used to identify different services or applications running on a networked device and are relevant for routing network traffic between devices.
Local port numbers are managed by the operating system and are not directly visible or relevant outside the local machine, whereas ports are used for communication between devices on a network.